updated version control scripts
