+++ src/Host.h
@@ -3,8 +3,8 @@
 
 #include "Parameters.h"
 #include "Mosquito.h"
-#include "array.h"
-#include "Sampler_v4.h"
+#include "array_v1.h"
+#include "Sampler_v5.h"
 
 #include <vector>
 #include <map>
